Toolbox/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
requests>=2.31.0
gspread>=6.0.0
google-auth>=2.0.0
